fix: only bump published date if revision changes
`gtfs_creator` always sends the revision IDs of all disruptions it is
publishing; Arrow in turn would always update the "published" timestamp
on all the corresponding disruptions, even if their published revisions
didn't change. This made the timestamp not that useful, since it would
always just be the timestamp of the last GTFS deploy for all (published)
disruptions.

This changes the logic to only update the "published" timestamp when the
published revision changes.