fix: calendar "today" highlight was in UTC
The start and end date of a disruption are ultimately just dates, with
no time or zone information, but since JavaScript lacks a "plain date"
type we parse them on the frontend as `Date` objects with the time part
set to 00:00:00 in UTC. To make FullCalendar display these correctly,
we set the calendar's `timeZone` to UTC. This had the side effect that
the "today" highlight followed the current UTC time, causing "today" to
advance to the next day too early (from our time zone's perspective).

FullCalendar also accepts event start/end values as ISO date strings,
which are assumed to be in the browser's local time zone. By doing so
and setting the calendar to use the `local` time zone (the default), we
can avoid dealing with times in the calendar at all, and "today" should
align with local time.