feat: disruptions table is server-rendered
This is primarily a port of the `DisruptionIndex` and `DisruptionTable`
components to Phoenix controller and view logic. The filters that can
be applied to the index are now expressed as query params, incidentally
making it possible to link to a specific filter state of the index. The
`DisruptionCalendar` wrapper for `FullCalendar` remains, and now accepts
a JSON:API document so the server can easily pass data to it.