feat: remove "ready" state from disruptions
This removes the feature where the latest revision of a disruption can
be marked as "ready", and in order to mark a revision as "published" it
must be earlier than or equal to the "ready" revision.

Since much of the React frontend will shortly be deleted, only minimal
changes are made there to disable the "readying" interface.