chore: remove SASS
SASS is a big dependency that complicates the build process. We'd like to use
esbuild or maybe PostCSS/tailwind in the future, and SASS isn't a well-
supported part of that process.

The big dependency here for us is bootstrap, which we're moving away from.
This PR generated a bootstrap.css by running our bootstrap and overrides
through sass, and checked that file in. There are instructions in _bootstrap.scss
about how to update that file if necessary.

As part of switching to CSS:
* I unnested the selectors
* I switched SASS variables to CSS custom properties.