feat: disruption details page is server-rendered
Since the "delete" button was part of `ViewDisruption` and was not
mentioned as a separate item in the migration task, its functionality
is also migrated here, with a small change: It redirects back to the
disruption details page rather than the index, giving a more explicit
confirmation of the action (the "This disruption is deleted" note).

Some basic controller tests are also added for the non-API disruptions
controller, to ensure everything works together in the absence of full
acceptance tests.