feat: disruption form is server-rendered
Since this form is highly "interactive" it is fundamentally kept as a
React component, but it now renders a set of hidden inputs that allow
the form to be submitted to the server using the standard browser
behavior, instead of calling a JSON:API endpoint. The form tag itself,
the submit/cancel buttons, and the display of any submission errors are
now handled by Phoenix (though as a slight UX regression, all errors
are combined into a flash message at the top of the page, rather than
being shown on the specific fields they apply to).

This incidentally changes the "new" form to allow creating a disruption
with no adjustments. This was something we wanted to do in the imminent
future anyway, and doing it now simplifies some aspects of this task. A
minor change is made to the table view so the "adjustments" cell is not
completely blank in this case, but we'll revisit this once we enable
disruptions having a "mode" or "line" independent of their adjustments.

The form should also now be more accessible, as it uses `label`s,
`aria-label`s, and `fieldset`s with `legend`s, instead of labeling
fields and sections using headers and bolded text.