refactor: create FullCalendar events on the server
`DisruptionCalendar` was the last place the frontend still used
JSON:API, as a means to get the disruption data from the server into
React to then generate the FullCalendar events. By instead generating
the events on the server and more-or-less rendering the `FullCalendar`
component directly, we can delete a large amount of bespoke JSON:API
code and even cut a dependency (`rrule`) from the frontend.

For the moment, the backend still needs JSON:API to support the index
endpoint used by `gtfs_creator` and the `copy_db` mix task.