refactor: use introspection to clone revisions
This avoids duplicating all the field names of revisions and their
associations, so we don't have to remember to also update the clone
function whenever a field is added or removed.

By creating a changeset for the clone instead of immediately saving it
to the database, we also avoid the need to wrap updates and deletes in
a transaction.

We still duplicate the names of associations — theoretically we could
also avoid this with introspection, but it's less clear that doing so
would be worth the complexity.