dev: fix missing `export` in `.envrc.example`
The example file didn't assign the env vars using `export`, which is
required to work with `direnv` and also to use the file as a standalone
shell script.

This also adds the `ARROW_DOMAIN` and `ARROW_API_KEY` values used by
`mix copy_db` to the example file, and changes the default domain to be
the production app, since this is the one we typically use.