feat: sort adjustment icons/events consistently
Following several changes to how we display disruption adjustments, the
sorting of adjustment icons and events was not consistent:

* In the table view, icons were sorted by adjustment label (a property
  which is no longer actually displayed in the table)

* In the calendar view, events were also sorted by adjustment label,
  though with no regard for mode/line

* On the detail page, icons were sorted by the order the adjustments
  happened to be returned from the database (no order was specified)

This change aligns these places on a single "adjustment kind order",
which is the same one established by the filter toggles (with calendar
events additionally being grouped by ROW approval status).

Since this order is now defined in a single place, this also makes it
easier to change in the future without forgetting something.