fix: handle empty all_calendar.txt in HASTUS import
If a HASTUS export had an empty `all_calendar.txt` file the import
process would not produce any services because the export parsing relied
on iterating over the services as listed in that `all_calendar.txt`
file.

To fix this we read the list of unique `service_id`s from the
`all_trips.txt` file instead and then "join" that with the values from
`all_calendar.txt`. If there is no entry in the `all_calendar.txt` file
for that service_id then we still include the service but without any
service dates. The user can then enter the service dates in the UI.

[🏹🐛 Empty all_calendar.txt leads to weirdness in HASTUS Import](https://app.asana.com/0/584764604969369/1209816734759554/f)