fix(navbar): resolve header jitter (#1213)
Problem:
The navbar tends to "jitter" (more precisely, there is layout shift) when switching pages on the new DisruptionsV2 home view. See the video attached in the Asana ticket I've linked.

Solution:
Set navbar buttons to a fixed width so that their size doesn't change whenever the text within them is rendered. Also revised some stylistic issues with buttons on other pages, and fixed a bug that was breaking sorting for the stops page.