fix: error when attempting to delete shapes that are used by shuttles
Shape deletion would sometimes leave things in a broken state. If a
shape was used by a shuttle the delete operation would first delete the
saved shape file and then fail to delete the database record because of
the foreign key constraint on `shuttle_routes`. This change refactors
that code to prevent this broken state and provide better feedback to
the user.

- Reorders the database and saved shape file deletion so that the
  database delete happens _before_ the saved shape file delete. This
  ensures that if a database operation fails the saved shape file isn't
  deleted (this is what was causing the 500 errors).
- Wraps the delete in the database and the delete of the saved shape in
  a database transaction to ensure that a failure to delete the remote
  shape does not delete the database record
- Adds a check to the `delete_shape` method to check for shuttles that
  are using the shape. Returns an error including a message that lists
  the shuttles using that shape if the shape is currently in use.
- Displays the new error message as an error flash if deleting a shape
  fails

Note: this does not fix any shapes that were broken by the old behavior.