feat: add download of list of trips/services with missing shapes
Updates the edit hastus export form to special case when a hastus export
has trips with invalid or missing shapes.

Also adds a new `send_download` function in `CoreComponents` which uses
a new client-side event handler to send data over the existing
LiveSocket connection and trigger a client-side download of that data.
This makes sending one-off files from a LiveView much easier than the
alternative of creating a separate controller to send the file.

[Asana task](https://app.asana.com/1/15492006741476/project/584764604969369/task/1210125477521799?focus=true)